// src/type/ztree.d.ts
  export interface ITreeNode {
    checked?: boolean;
    children?: Array<ITreeNode>;
    chkDisabled?: boolean;
    click?: CallBackOnFn;
    getCheckStatus?: () => object;
    getIndex?: () => number;
    getNextNode?: () => ITreeNode;
    getParentNode?: () => ITreeNode;
    getPath?: () => Array<ITreeNode>;
    getPrevNode?: () => ITreeNode;
    halfCheck?: boolean;
    icon?: string;
    iconClose?: string;
    iconOpen?: string;
    iconSkin?: string;
    isHidden?: boolean;
    isParent?: boolean;
    name?: string;
    nocheck?: boolean;
    open?: boolean;
    target?: string;
    url?: string;
    check_Child_State?: number;
    check_Focus?: boolean;
    checkedOld?: boolean;
    editNameFlag?: boolean;
    isAjaxing?: boolean;
    isFirstNode?: boolean;
    isHover?: boolean;
    isLastNode?: boolean;
    level?: number;
    parentTId?: string;
    tId?: string;
    [key: string]: any;
  }
// ...
  type CallBackOnFn = (event: Event, treeId: string, treeNode: ITreeNode) => void | Promise<void>;